% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_ana.R
\name{label_ana}
\alias{label_ana}
\title{Use ANA file to label raw data}
\usage{
label_ana(data, ana)
}
\arguments{
\item{data}{Dataframe containing columns for time and volts}

\item{ana}{A dataframe for an ANA file.}
}
\value{
A tibble object with three rows is returned: time, volts, waveform.
}
\description{
The function label_ana combines raw data and an
ANA annotation together to return a labeled dataframe.
}
\details{
The raw data and ana must be for the same dataset, although the
names of the objects do not need to match in any way. The following code
number to character title conversion is followed:
1 = non-probing
2 = C
4 = E1
5 = E2
7 = G
8 = pd
9 = pd1
10 = pd2
99 = end
If the ANA labels do not come in that exact code matching, then the labeling
will be incorrect.
}
\seealso{
\code{\link{read_epg}} to read in the raw data and ANA files.
}
