% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfreqs.R
\name{mainfreqs}
\alias{mainfreqs}
\title{Find the main frequencies in a sample using the Fast Fourier Transform}
\usage{
mainfreqs(data, end = sample_rate/2, pre_fft = FALSE,
window = TRUE, filter = 0.25, sample_rate = 100)
}
\arguments{
\item{data}{Either a voltage time series dataframe or a dataframe of Fourier
transformed data with columns for frequency and amplitude. If the latter,
change pre_fft to TRUE.}

\item{end}{Numeric indicating number of frequencies to display, e.g. end = 10
will be top 10 frequencies. Default is all (half sample_rate)}

\item{pre_fft}{TRUE/FALSE indicating whether or not the data being fed has been
Fourier Transformed,}

\item{window}{Default true will window the data before transforming.}

\item{filter}{High pass filter indicating threshold frequency above which
will be kept, and below which will be filtered out}

\item{sample_rate}{The frequency of collection in Hz. Default is 100 Hz}
}
\value{
A dataframe containing the main frequencies with 2 columns is returned:
frequency and amplitude.
}
\description{
The function mainfreqs finds the main frequencies in voltage time-
series data by using the Fast Fourier Transform.
}
\details{
Main frequencies are determined as those having the
largest amplitude. Frequencies are aggregated such that only the highest in each
integer bin are kept and compared - e.g. if the data is 5.1 Hz, 5.2 Hz, only the
largest will be kept for the 5.x Hz entry. Therefore if there are multiple main
frequencies in the same bin, only the largest one will the kept and the others
will be discarded.
}
\seealso{
Other frequency related functions: 
\code{\link{plot_fbox}()},
\code{\link{plot_fft}()},
\code{\link{single_fft}()},
\code{\link{topfreq}()},
\code{\link{wave_topfreq}()}
}
\concept{frequency related functions}
