% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fbar.R
\name{plot_fbar}
\alias{plot_fbar}
\title{Bar Chart of Frequencies}
\usage{
plot_fbar(data, waveforms = c("A", "C", "E1", "E2", "G", "pd1",
"pd2", "pd"))
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}

\item{waveforms}{A list of waveforms to include in the plot. Default is all.}
}
\value{
A bar chart of frequencies by waveform
}
\description{
The function plot_fbar shows the frequencies present in a dataset
plotted in a bar chart by waveform.
}
\details{
The function plot_fbar is designed to allow one to see the frequencies
and the variability of frequencies for waveforms in a dataset.
}
\seealso{
Other frequency related functions: 
\code{\link{mainfreqs}()},
\code{\link{plot_fft}()},
\code{\link{single_fft}()},
\code{\link{topfreq}()},
\code{\link{wave_topfreq}()}
}
\concept{frequency related functions}
