% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fft.R
\name{plot_fft}
\alias{plot_fft}
\title{Plot of amplitude vs frequency for transformed voltage data}
\usage{
plot_fft(data, end = NULL, include_dc = FALSE, pre_fft = FALSE,
window = TRUE, filter = 0.25, sample_rate = 100)
}
\arguments{
\item{data}{Either a voltage time series dataframe or a dataframe of Fourier
transformed data with columns for frequency and amplitude. If the latter,
change pre_fft to TRUE.}

\item{end}{Default is NULL to plot all data. Otherwise, insert numeric value
corresponding to the maximum frequency you would like to plot. This does
not affect the frequency values themselves, it merely selects the portion
of the graph to view.}

\item{include_dc}{Default true, will include DC component (amplitude at 0 Hz)}

\item{pre_fft}{TRUE/FALSE indicating whether or not the data being fed has been
Fourier Transformed,}

\item{window}{Default true will window the data before transforming.}

\item{filter}{High pass filter indicating threshold frequency above which
will be kept, and below which will be filtered out}

\item{sample_rate}{The frequency of collection in Hz. Default is 100 Hz}
}
\value{
A plot showing amplitude vs frequency of the Fourier Transformed data.
}
\description{
The function plot_fft allows one to visualize the results
of a Fourier Transform and check for peaks - main constituent frequencies.
}
\details{
If window = TRUE is chosen, then a Hann (Hanning) window is applied
to the data for the transform. Additionally, window = TRUE will select a
2^10 = 1024 point data.frame from the center of the data in order to get
a 'center cut' of the data and avoid irregularities around the edges.
}
\seealso{
Other frequency related functions: 
\code{\link{mainfreqs}()},
\code{\link{plot_fbar}()},
\code{\link{single_fft}()},
\code{\link{topfreq}()},
\code{\link{wave_topfreq}()}
}
\concept{frequency related functions}
