% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie.R
\name{plot_pie}
\alias{plot_pie}
\title{Pie chart of labelled EPG data}
\usage{
plot_pie(data, pietype = c("time", "count"), sample_rate = 100)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}

\item{pietype}{Select the type of pie chart to display. Default is time - a pie
chart showing the percent of time spent in each waveform is displayed. If "count"
is selected, then the number of instances for each waveform type is displayed - e.g.
if there are 20 distince waveforms, and 5 are of type C, then C will be 5/20 or
25\%.}

\item{sample_rate}{The frequency of collection in Hz. Default is 100 Hz}
}
\value{
A pie chart of EPG waveforms in a dataset.
}
\description{
The function plot_pie allows one to visualize EPG waveforms in a
pie chart format.
}
\details{
The function plot_pie is designed to help visualize EPG waveform data.
}
\seealso{
Other waveform functions: 
\code{\link{plot_wave}()},
\code{\link{wave_duration}()},
\code{\link{wave_extract}()},
\code{\link{wave_topfreq}()}
}
\concept{waveform functions}
