% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vts.R
\name{plot_vts}
\alias{plot_vts}
\title{Time series plot of voltage}
\usage{
plot_vts(data, aggregate = TRUE, interval = 100)
}
\arguments{
\item{data}{Dataframe containing columns for time and volts}

\item{aggregate}{TRUE/FALSE indicated whether to aggregate the data and
reduce the number of points by plotting only the mean over a set interval.}

\item{interval}{Integer indicating the number of points to aggregate
together if aggregate = TRUE.}
}
\value{
A voltage time series plot.
}
\description{
The function plot_vts plots voltage vs time.
}
\details{
If a large amount of data is fed in, it is recommended to aggregate
because plotting an excess of points is time intensive.
}
