% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epg.R
\name{read_epg}
\alias{read_epg}
\title{Read in EPG data files}
\usage{
read_epg(data, extension = c("txt", "ANA", "csv"))
}
\arguments{
\item{data}{As with other read in functions, insert the file
path for the data file here.}

\item{extension}{The file extension for the raw data that is to be read. The
options are txt, ANA, and csv. If no extension is indicated,
txt will be chosen by defaults.}
}
\value{
A tibble object corresponding to the type of data input is returned.
}
\description{
The function read_epg data reads in EPG data.
}
\details{
Three file formats are supported: raw data text files with format
time;volts, ANA annotation files with columns in order waveform, time,
and volts (they do not need to be named), and lastly csv files of combined
raw data and annotations with columns labeled time, volts, and waveform.
}
