% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_hr}
\alias{select_hr}
\title{Select a single hour out of a larger dataset.}
\usage{
select_hr(data, hour, sample_rate = 100)
}
\arguments{
\item{data}{May be raw voltage time series data or labelled time series data.}

\item{hour}{Integer indicating which hour of data to subset - e.g. 2 will select
the second hour of data.}

\item{sample_rate}{The frequency of collection in Hz. Default is 100 Hz}
}
\value{
A dataframe subsetting the input data to the specified hour is returned.
}
\description{
The function select_hr subsets a user specified hour
from a larger dataset.
}
\details{
If the selected hour is outside of the data range, or the data has gaps,
or the hour is incomplete (as is possible for the last hour), the function may
function incorrectly.
}
