% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_fft.R
\name{single_fft}
\alias{single_fft}
\title{Return the frequency and amplitude using the Fast Fourier Transform}
\usage{
single_fft(data, window = TRUE, include_dc = TRUE, filter = 0.25,
sample_rate = 100)
}
\arguments{
\item{data}{Dataframe containing columns for time and volts}

\item{window}{Default true will window the data before transforming.}

\item{include_dc}{Default true, will include DC component (amplitude at 0 Hz)}

\item{filter}{High pass filter indicating threshold frequency above which
will be kept, and below which will be filtered out}

\item{sample_rate}{The frequency of collection in Hz. Default is 100 Hz}
}
\value{
A tibble object with two columns: Frequency and Amplitude. It is
single-sided, and will select up to the Nyquist (sample_rate/2)
}
\description{
The function, single_fft, runs the R implementation of the
Fast Fourier Transform, fft, and returns a dataframe containing the single-
sided FFT with columns for frequency and amplitude.
}
\details{
If window = TRUE is chosen, then a Hann (Hanning) window is applied
to the data for the transform. Additionally, window = TRUE will select a
2^10 = 1024 point data.frame from the center of the data in order to get
a 'center cut' of the data and avoid irregularities around the edges.
}
\seealso{
Other frequency related functions: 
\code{\link{mainfreqs}()},
\code{\link{plot_fbar}()},
\code{\link{plot_fft}()},
\code{\link{topfreq}()},
\code{\link{wave_topfreq}()}
}
\concept{frequency related functions}
