% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topfreq.R
\name{topfreq}
\alias{topfreq}
\title{Calculate the single main frequency for a voltage time series}
\usage{
topfreq(data, window = TRUE, filter = 0.25, pre_fft = FALSE)
}
\arguments{
\item{data}{Either a voltage time series dataframe or a dataframe of Fourier
transformed data with columns for frequency and amplitude. If the latter,
change pre_fft to TRUE.}

\item{window}{Default true will window the data before transforming.}

\item{filter}{High pass filter indicating threshold frequency above which
will be kept, and below which will be filtered out}

\item{pre_fft}{TRUE/FALSE indicating whether or not the data being fed has been
Fourier Transformed,}
}
\value{
The single main frequency and its amplitude is returned as a tibble object.
}
\description{
The function topfreq calculates the single main frequency
for voltage time series data using the Fast Fourier Transform.
}
\details{
The main frequency is defined as that which has the largest amplitude.
}
\seealso{
Other frequency related functions: 
\code{\link{mainfreqs}()},
\code{\link{plot_fft}()},
\code{\link{single_fft}()},
\code{\link{wave_topfreq}()}
}
\concept{frequency related functions}
