% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_extract.R
\name{wave_extract}
\alias{wave_extract}
\title{Extract a specified waveform from a labelled dataset}
\usage{
wave_extract(data, wave, number = 1)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}

\item{wave}{Desired type of waveform, e.g. pd1}

\item{number}{Integer denoting which instance of the specified waveform type to
select - e.g. number = 1 will select the first instance of waveform type specified
by wave parameter.}
}
\value{
A list containing 2 entries is returned: entry one is a tibble of
time, volts, waveform is returned containing just the specified
waveform instance. List entry 2 is the number/count of that type of waveform -
e.g. if wave = "pd2" and there are 12 pd2 instances in the dataset, this
entry will return 12.
}
\description{
The function wave_extract subsets data and returns the specified
waveform.
}
\details{
If the wave parameter input does not exactly match a waveform label
in the data, the function will not work.
}
\seealso{
Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_duration}()},
\code{\link{wave_number}()},
\code{\link{wave_topfreq}()},
\code{\link{wave_volts}()}
}
\concept{waveform functions}
