% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_occurrence.R
\name{wave_occurrence}
\alias{wave_occurrence}
\title{Calculate the Number of each waveform}
\usage{
wave_occurrence(data)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}
}
\value{
A tibble object containing a row per waveform instance and two
columns, waveform and occurrence, is returned.
}
\description{
The function wave_occurrence calculates the number of each waveform
type.
}
\details{
Labelled data is grouped by each waveform instance and the
unique instances of each waveform type are counted. The waveform type
'pd' independent of any splitting into subforms is calculated at the end.
As a result, the pd calculations  will appear all at the end of the table.
The subforms, pd1 and pd2, will appear in sequence.
}
\seealso{
Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_duration}()},
\code{\link{wave_extract}()},
\code{\link{wave_topfreq}()},
\code{\link{wave_volts}()}
}
\concept{waveform functions}
