% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_topfreq.R
\name{wave_topfreq}
\alias{wave_topfreq}
\title{Calculate and Extract the top frequency for each waveform instance}
\usage{
wave_topfreq(data)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}
}
\value{
A tibble object containing a row per waveform instance and two
columns, waveform and frequency, is returned.
}
\description{
The function wave_topfreq calculates the main/top frequency for
each waveform instance in labelled data.
}
\details{
Labelled data is grouped by each waveform instance and the
Fourier transform is then used to extract the top frequency via the
epgminer function topfreq. The waveform type 'pd' independent of any
splitting into subforms is calculated at the end. As a result, the
pd calculations will appear all at the end of the table. The subforms,
pd1 and pd2, will appear in sequence. Frequency is reported in Hz.
}
\seealso{
Other frequency related functions: 
\code{\link{mainfreqs}()},
\code{\link{plot_fbox}()},
\code{\link{plot_fft}()},
\code{\link{single_fft}()},
\code{\link{topfreq}()}

Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_duration}()},
\code{\link{wave_extract}()},
\code{\link{wave_number}()},
\code{\link{wave_volts}()}
}
\concept{frequency related functions}
\concept{waveform functions}
