% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_duration.R
\name{wave_duration}
\alias{wave_duration}
\title{Calculate and Extract the duration for each waveform instance}
\usage{
wave_duration(data)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}
}
\value{
A tibble object containing a row per waveform instance and two
columns, waveform and duration, is returned.
}
\description{
The function wave_duration calculates the duration for
each waveform instance in labelled data.
}
\details{
Labelled data is grouped by each waveform instance and the duration
is then calculated. The overall feeding duration is displayed in the top row
as "Feeding" for simplicity. The waveform type 'pd' independent of any splitting
into subforms is calculated at the end. As a result, the pd calculations
will appear all at the end of the table. The subforms, pd1 and pd2,
will appear in sequence. Duration is reported in seconds.
}
\seealso{
Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_extract}()},
\code{\link{wave_number}()},
\code{\link{wave_topfreq}()}
}
\concept{waveform functions}
