% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_duration.R
\name{wave_duration}
\alias{wave_duration}
\title{Calculate and Extract the duration for each waveform instance}
\usage{
wave_duration(data)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}
}
\value{
A tibble object containing a row per waveform instance and two
columns, waveform and duration, is returned.
}
\description{
The function wave_duration calculates the duration for
each waveform instance in labelled data.
}
\details{
Labelled data is grouped by each waveform instance and the duration
is then calculated.
}
\seealso{
Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_extract}()},
\code{\link{wave_topfreq}()}
}
\concept{waveform functions}
