% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_volts.R
\name{wave_volts}
\alias{wave_volts}
\title{Calculate voltage metrics for each waveform}
\usage{
wave_volts(data)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}
}
\value{
A tibble object containing a row per waveform instance and four
columns is returned: waveform, mean_volts, sd_volts, relative_amplitude.
}
\description{
The function wave_volts calculates the voltage metrics
for each waveform instance
}
\details{
Labeled data is grouped by each waveform instance and the following three
voltage metrics are returned: mean, SD, and relative amplitude.
The SD is the mathematical SD of volts. Mean is first corrected for by subtracting
the starting voltage, then the mean (adjusted) voltage is returned.
The waveform type 'pd' without any splitting into subforms is calculated separately.
As a result, the pd calculations will appear all at the end of the table.
The subforms, pd1 and pd2, will appear in sequence.

Relative amplitude is a percent calculated with the following formula:
Relative amplitude = |(mean of waveform - mean of nonprobing)|x100/5
}
\seealso{
Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_duration}()},
\code{\link{wave_extract}()},
\code{\link{wave_occurrence}()},
\code{\link{wave_topfreq}()}
}
\concept{waveform functions}
