% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_volts.R
\name{wave_volts}
\alias{wave_volts}
\title{Calculate voltage metrics for each waveform}
\usage{
wave_volts(data)
}
\arguments{
\item{data}{A dataframe containing three columns: time, volts, waveform.}
}
\value{
A tibble object containing a row per waveform instance and four
columns is returned: waveform, median_volts, sd_volts, amplitude_volts.
}
\description{
The function wave_volts calculates the voltage metrics
for each waveform instance
}
\details{
Labeled data is grouped by each waveform instance and the following three
voltage metrics are returned: median, SD, and amplitude.
The amplitude is defined as the range of voltage values while the SD is the
mathematical SD. Median is first corrected for by subtracting the starting
voltage, then the median relative (adjusted) voltage is returned.
}
\seealso{
Other waveform functions: 
\code{\link{plot_pie}()},
\code{\link{plot_wave}()},
\code{\link{wave_duration}()},
\code{\link{wave_extract}()},
\code{\link{wave_number}()},
\code{\link{wave_topfreq}()}
}
\concept{waveform functions}
